refactor(core): keep the renderer tool vocabulary in one place
The vocabulary existed three times: the enum and its fold in the server, a
separate 20-arm match in the store for names read back from the journal, and
ChatToolActivitySnapshot's own field, typed `&'static str`. The store copy had
already drifted once — `exec` was missing from it, so a command read as "Ran a
command" while streaming and "Used a tool" after a reload, with its own command
card still visible underneath.

Move the enum and one fold into openwave-core, where the store can see it, and
type the snapshot field as it. The store's match is deleted rather than checked
against the enum: one mapping cannot disagree with itself.

The fold uses literals where it used to reference the server's tool-name
constants, which this crate cannot see. `source_tools_use_fixed_renderer_names`
already holds those constants to these spellings, so the link is kept where it
was rather than reintroduced here.

Typing the field also closes the first hazard in #548. As `&'static str` it
generated as `string`, which compiles on both sides while silently dropping the
allowlist that the renderer's copy and icon tables are keyed on — the
`Record<RendererToolName, _>` totality that makes a tool without presentation a
compile error. It now generates as the union, and the snapshot is generated
rather than hand-written, so the two cannot drift.

The 20-name round-trip test moves with the function it covered. Its input is the
registered tool names, which are not derivable from the enum: the fold exists
because the two vocabularies are different sets that happen to agree on
spelling.
